backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
argcomplete==3.6.2
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.10.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.4.0
autocommand==2.2.2
babel==2.17.0
beautifulsoup4==4.14.2
bleach==6.3.0
btrfsutil==6.17.1
build==1.3.0
cattrs==25.3.0
certifi==2025.11.12
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.2.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.3
cupshelpers==1.0
curl_cffi==0.13.0
cycler==0.12.1
dbus-python==1.4.0
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
Django==5.1.14
django-allauth==65.13.0
django-cors-headers==4.8.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnspython==2.8.0
eventlet==0.40.3
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.20.0
fonttools==4.60.1
frozendict==2.4.6
gevent==25.9.1
Glances==4.4.1
greenlet==3.2.4
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.0
iniconfig==2.1.0
installer==0.7.0
ipykernel==7.1.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jaraco.collections==5.1.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jaraco.text==4.0.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
json5==0.12.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyterlab==4.5.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
kiwisolver==1.4.7
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
mistune==3.1.4
more-itertools==10.8.0
multitasking==0.0.12
mysqlclient==2.2.7
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest_asyncio==1.6.0
netsnmp-python==1.0a1
nftables==0.1
notebook==7.5.0
notebook_shim==0.2.4
numpy==2.3.5
orjson==3.11.4
packaging==25.0
pandas==2.3.1
pandocfilters==1.5.1
parso==0.8.5
peewee==3.18.3
pexpect==4.9.0
phonenumbers==9.0.17
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
pooch==1.8.2
prometheus_client==0.23.1
prompt_toolkit==3.0.52
protobuf==6.33.1
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyaml==24.12.0
pycairo==1.29.0
pycparser==2.23
pycups==2.0.4
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
PyGObject==3.54.5
PyJWT==2.10.1
pyparsing==3.2.5
pyproject_hooks==1.2.0
PyQt5==5.15.11
PyQt5_sip==12.17.1
PyQt6==6.10.0
PyQt6_sip==13.10.2
pytest==8.4.2
python-dateutil==2.9.0
python-dotenv==1.2.1
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
QTermWidget==2.3.0
referencing==0.37.0
Reflector==2023.6.28.0.36.1
reportlab==4.2.2
requests==2.32.5
requests-cache==1.2.1
requests-mock==1.12.1
responses==0.25.8
rich==14.2.0
rpds-py==0.29.0
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.15
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==80.9.0
shtab==1.8.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
sqlparse==0.5.3
stack_data==0.6.3
TBB==0.2
termcolor==3.2.0
threadpoolctl==3.5.0
time-machine==2.19.0
timeout-decorator==0.5.0
tinycss2==1.4.0
tldr==3.4.3
tornado==6.5.2
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
ujson==5.11.0
url-normalize==2.2.1
urllib3==2.5.0
wcwidth==0.2.14
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
wheel==0.45.1
yfinance==0.2.65
zipp==3.21.0
zope.event==6.1
zope.interface==8.1.1
